move type helper to typings
