+++ lib/src/_shared/PickerToolbar.tsx
@@ -5,7 +5,7 @@ import Toolbar, { ToolbarProps } from '@material-ui/core/Toolbar';
 import classnames from 'classnames';
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
-import { ExtendMui } from '../_helpers/types';
+import { ExtendMui } from '../typings/extendMui';
 
 export interface PickerToolbarProps
   extends ExtendMui<ToolbarProps>,
